refactor(picker): refactor picker
