+++ packages/zarm/src/date-picker-view/constant.ts
@@ -1,0 +1,6 @@
+export const DATETIME = 'datetime';
+export const DATE = 'date';
+export const TIME = 'time';
+export const MONTH = 'month';
+export const YEAR = 'year';
+export const ONE_DAY = 24 * 60 * 60 * 1000;
