+++ packages/zarm/src/date-select/index.ts
@@ -1,5 +1,5 @@
 import DateSelect from './DateSelect';
 
-export type { DateSelectProps } from './DateSelect';
+export type { DateSelectProps, BaseDateSelectProps } from './interface';
 
 export default DateSelect;
