+++ packages/zarm/src/picker-view/index.ts
@@ -1,6 +1,5 @@
 import PickerView from './PickerView';
 
-export type { PickerDataSource } from './interface';
-export type { PickerViewProps } from './PickerView';
+export type { PickerViewProps, PickerDataSource, BasePickerViewProps } from './interface';
 
 export default PickerView;
