+++ packages/zarm/src/picker/index.ts
@@ -1,5 +1,5 @@
 import Picker from './Picker';
 
-export type { PickerProps } from './Picker';
+export type { PickerProps, BasePickerProps } from './interface';
 
 export default Picker;
